Converted to int as status code#

+++ internal/core/status/models.go
@@ -12,5 +12,5 @@ type InfoStatus struct {
 	CountriesNowAPI  int    `json:"countriesnowapi"`
 	RestCountriesAPI int    `json:"restcountriesapi"`
 	Version          string `json:"version"`
-	Uptime           int64  `json:"uptime"`
+	Uptime           int    `json:"uptime"`
 }
